# lizmap_server/api/openapi.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/api/v1/projects/description": {
            "get": {
                "summary": "Project description",
                "description": "Return a project description\n",
                "parameters": [
                    {
                        "in": "query",
                        "name": "map",
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "description": "The project path"
                    }
                ],
                "tags": [
                    "project"
                ],
                "responses": {
                    "200": {
                        "description": "The project's description\n",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/definitions/ProjectDescription"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "The project does not exists\n",
                        "content": {
                            "application/json": {
                                "schema": null
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/projects/layers/{Id}": {
            "get": {
                "summary": "Project's layer details",
                "description": "Return the project layers details for layer {Id}\n",
                "parameters": [
                    {
                        "in": "path",
                        "name": "Id",
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "description": "Project's layer id"
                    },
                    {
                        "in": "query",
                        "name": "p",
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "description": "The project path"
                    }
                ],
                "tags": [
                    "project"
                ],
                "responses": {
                    "200": {
                        "description": "The project's layer details\n",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/definitions/LayerDetails"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "The layer does not exists\n",
                        "content": {
                            "application/json": {
                                "schema": null
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/projects/layouts/": {
            "get": {
                "summary": "Project's layout summary",
                "description": "Return the project layout summaries\n",
                "parameters": [
                    {
                        "in": "query",
                        "name": "p",
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "description": "The project path"
                    }
                ],
                "tags": [
                    "layout"
                ],
                "responses": {
                    "200": {
                        "description": "The `roject's layout summaries\n",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/definitions/LayoutSummaries"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "The project does not exists\n",
                        "content": {
                            "application/json": {
                                "schema": null
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/projects/layouts/{Name}": {
            "get": {
                "summary": "Project's layout details",
                "description": "Return the project's layout details for layout {Id}\n",
                "parameters": [
                    {
                        "in": "path",
                        "name": "Name",
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "description": "Project's layout name"
                    },
                    {
                        "in": "query",
                        "name": "p",
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "description": "The project path"
                    }
                ],
                "tags": [
                    "layout"
                ],
                "responses": {
                    "200": {
                        "description": "The project's layer details\n",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/definitions/LayoutDescription"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "The layout or the project does not exists\n",
                        "content": {
                            "application/json": {
                                "schema": null
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "CrsModel": {
            "additionalProperties": true,
            "description": "CRS",
            "properties": {
                "authId": {
                    "description": "The authority Id of the projection",
                    "type": "string"
                },
                "proj": {
                    "description": "The proj string for the crs",
                    "type": "string"
                },
                "srid": {
                    "description": "The Postgis SRID",
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                }
            },
            "required": [
                "authId",
                "proj",
                "srid",
                "description"
            ],
            "type": "object"
        },
...
        "GuiProperties": {
            "additionalProperties": true,
            "description": "Gui properties",
            "properties": {
                "backgroundColor": {
                    "format": "color",
                    "type": "string"
                },
                "selectionColor": {
                    "format": "color",
                    "type": "string"
                }
            },
            "type": "object"
        },
        "LayerDescription": {
            "additionalProperties": true,
            "description": "Layer description",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "abstract": {
                    "type": "string"
                },
                "embedded": {
                    "type": "boolean"
                },
                "provider": {
                    "type": "string"
                },
                "datasource": {
                    "type": "string"
                },
                "opacity": {
                    "type": "number"
                },
                "srs": {
                    "$ref": "#/definitions/CrsModel"
                },
                "styleManager": {
                    "$ref": "#/definitions/StyleManager"
                },
                "keywords": {
                    "type": "string"
                },
                "attributeAliases": {
                    "additionalProperties": {
                        "type": "string"
                    },
                    "type": "object"
                },
                "excludeAttributeWfs": {
                    "items": {
                        "type": "string"
                    },
                    "type": "array"
                },
                "excludeAttributeWms": {
                    "items": {
                        "type": "string"
                    },
                    "type": "array"
                },
                "displayExpression": {
                    "type": "string"
                },
                "vectorJoins": {
                    "items": {
                        "$ref": "#/definitions/VectorLayerJoinInfo"
                    },
                    "type": "array"
                },
                "featureRenderer": {
                    "$ref": "#/definitions/FeatureRenderer"
                }
            },
            "required": [
                "id",
                "name",
                "type",
                "abstract",
                "embedded",
                "provider",
                "datasource",
                "opacity",
                "styleManager",
                "keywords",
                "attributeAliases",
                "excludeAttributeWfs",
                "excludeAttributeWms",
                "vectorJoins"
            ],
            "type": "object"
        },
...
        "LayerTreeRoot": {
            "additionalProperties": true,
            "properties": {
                "customProperties": {
                    "items": {
                        "type": "string"
                    },
                    "type": "array"
                },
                "customOrderEnabled": {
                    "type": "boolean"
                },
                "customOrderItems": {
                    "items": {
                        "type": "string"
                    },
                    "type": "array"
                },
                "items": {
                    "items": {
                        "anyOf": [
                            {
                                "$ref": "#/definitions/LayerTreeLayer"
                            },
                            {
                                "$ref": "#/definitions/LayerTreeGroup"
                            }
                        ]
                    },
                    "type": "array"
                }
            },
            "required": [
                "customProperties",
                "customOrderEnabled",
                "customOrderItems",
                "items"
            ],
            "type": "object"
        },
...
        "LayerVisibilityPreset": {
            "additionalProperties": true,
            "properties": {
                "name": {
                    "type": "string"
                },
                "layers": {
                    "items": {
                        "$ref": "#/definitions/LayerVisibility"
                    },
                    "type": "array"
                },
                "checkedGroupNodes": {
                    "items": {
                        "type": "string"
                    },
                    "type": "array",
                    "uniqueItems": true
                },
                "expandedGroupNodes": {
                    "items": {
                        "type": "string"
                    },
                    "type": "array",
                    "uniqueItems": true
                }
            },
            "required": [
                "name",
                "layers",
                "checkedGroupNodes",
                "expandedGroupNodes"
            ],
            "type": "object"
        },
        "OWSProperties": {
            "additionalProperties": true,
            "description": "Ows properties",
            "properties": {
                "owsServiceTitle": {
                    "type": "string"
                },
                "owsServiceAbstract": {
                    "type": "string"
                },
                "owsServiceKeywords": {
                    "items": {
                        "type": "string"
                    },
                    "type": "array"
                },
                "owsOnlineResource": {
                    "type": "string"
                },
                "owsServiceContactMail": {
                    "type": "string"
                },
                "owsServiceContactOrganization": {
                    "type": "string"
                },
                "owsServiceContactPerson": {
                    "type": "string"
                },
                "owsServiceContactPhone": {
                    "type": "string"
                },
                "wwsMaxHeight": {
                    "type": "integer"
                },
                "wmsMaxWidth": {
                    "type": "integer"
                },
                "wmsExtent": {
                    "oneOf": [
                        {
                            "items": {
                                "type": "number"
                            },
                            "maxItems": 4,
                            "minItems": 4,
                            "type": "array"
                        },
                        {
                            "items": {
                                "type": "number"
                            },
                            "maxItems": 6,
                            "minItems": 6,
                            "type": "array"
                        }
                    ]
                },
                "wmsRestrictedLayers": {
                    "items": {
                        "type": "string"
                    },
                    "type": "array"
                },
                "wmsMaxAtlasFeatures": {
                    "type": "integer"
                },
                "wmsRestrictedComposers": {
                    "items": {
                        "type": "string"
                    },
                    "type": "array"
                },
                "wmsUseLayerIds": {
                    "type": "boolean"
                },
                "wmsFeatureInfoAddWktGeometry": {
                    "type": "boolean"
                },
                "wfsLayersIds": {
                    "items": {
                        "type": "string"
                    },
                    "type": "array"
                }
            },
            "required": [
                "owsServiceTitle",
                "owsServiceAbstract",
                "owsServiceKeywords",
                "owsOnlineResource",
                "owsServiceContactMail",
                "owsServiceContactOrganization",
                "owsServiceContactPerson",
                "owsServiceContactPhone",
                "wwsMaxHeight",
                "wmsMaxWidth",
                "wmsRestrictedLayers",
                "wmsMaxAtlasFeatures",
                "wmsRestrictedComposers",
                "wmsUseLayerIds",
                "wmsFeatureInfoAddWktGeometry",
                "wfsLayersIds"
            ],
            "type": "object"
        },
...
        "ProjectDescription": {
            "additionalProperties": true,
            "description": "Project description",
            "properties": {
                "title": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                },
                "saveDateTime": {
                    "format": "date-time",
                    "type": "string"
                },
                "projectCrs": {
                    "$ref": "#/definitions/CrsModel"
                },
                "owsProperties": {
                    "$ref": "#/definitions/OWSProperties"
                },
                "guiProperties": {
                    "$ref": "#/definitions/GuiProperties"
                },
                "layerTreeRoot": {
                    "$ref": "#/definitions/LayerTreeRoot"
                },
                "visibilityPresets": {
                    "items": {
                        "$ref": "#/definitions/LayerVisibilityPreset"
                    },
                    "type": "array"
                },
                "customVariables": {
                    "additionalProperties": {
                        "type": "string"
                    },
                    "type": "object"
                },
                "layers": {
                    "additionalProperties": {
                        "$ref": "#/definitions/LayerDescription"
                    },
                    "type": "object"
                }
            },
            "required": [
                "title",
                "version",
                "layerTreeRoot",
                "visibilityPresets",
                "customVariables",
                "layers"
            ],
            "title": "ProjectDescription",
            "type": "object"
        },
...
        "RasterLayerDetails": {
            "additionalProperties": true,
            "properties": {
                "provider": {
                    "title": "Provider",
                    "type": "string"
                },
                "type": {
                    "const": "raster",
                    "default": "raster",
                    "title": "Type",
                    "type": "string"
                }
            },
            "required": [
                "provider"
            ],
            "title": "RasterLayerDetails",
            "type": "object"
        },
        "VectorLayerDetails": {
            "additionalProperties": true,
            "properties": {
                "provider": {
                    "title": "Provider",
                    "type": "string"
                },
                "type": {
                    "const": "vector",
                    "default": "vector",
                    "title": "Type",
                    "type": "string"
                },
                "attributeTableConfig": {
                    "$ref": "#/definitions/AttributeTableConfig",
                    "default": {
                        "columns": []
                    }
                },
                "editableFields": {
                    "additionalProperties": {
                        "type": "boolean"
                    },
                    "title": "Editablefields",
                    "type": "object"
                },
                "fields": {
                    "additionalProperties": {
                        "$ref": "#/definitions/FieldConfiguration"
                    },
                    "title": "Fields",
                    "type": "object"
                },
                "defaults": {
                    "additionalProperties": {
                        "$ref": "#/definitions/FieldDefault"
                    },
                    "title": "Defaults",
                    "type": "object"
                },
                "constraints": {
                    "additionalProperties": {
                        "$ref": "#/definitions/FieldConstraint"
                    },
                    "title": "Constraints",
                    "type": "object"
                },
                "constraintExpressions": {
                    "additionalProperties": {
                        "$ref": "#/definitions/FieldConstraintExpression"
                    },
                    "title": "Constraintexpressions",
                    "type": "object"
                }
            },
            "required": [
                "provider",
                "editableFields",
                "fields",
                "defaults",
                "constraints",
                "constraintExpressions"
            ],
            "title": "VectorLayerDetails",
            "type": "object"
        },
        "LayerDetails": {
            "anyOf": [
                {
                    "$ref": "#/definitions/VectorLayerDetails"
                },
                {
                    "$ref": "#/definitions/RasterLayerDetails"
                }
            ]
        },
...
        "LayoutPage": {
            "additionalProperties": true,
            "properties": {
                "width": {
                    "type": "number"
                },
                "heigth": {
                    "type": "number"
                },
                "units": {
                    "type": "string"
                }
            },
            "required": [
                "width",
                "heigth",
                "units"
            ],
            "type": "object"
        },
        "LayoutDescription": {
            "additionalProperties": true,
            "properties": {
                "name": {
                    "type": "string"
                },
                "atlasEnabled": {
                    "type": "boolean"
                },
                "atlasCoverageLayer": {
                    "type": "string"
                },
                "pages": {
                    "items": {
                        "$ref": "#/definitions/LayoutPage"
                    },
                    "type": "array"
                },
                "maps": {
                    "items": {
                        "$ref": "#/definitions/LayoutMap"
                    },
                    "type": "array"
                },
                "labels": {
                    "items": {
                        "$ref": "#/definitions/LayoutLabel"
                    },
                    "type": "array"
                }
            },
            "required": [
                "name",
                "atlasEnabled",
                "pages",
                "maps",
                "labels"
            ],
            "title": "LayoutDescription",
            "type": "object"
        },
        "LayoutSummary": {
            "additionalProperties": true,
            "properties": {
                "name": {
                    "type": "string"
                },
                "atlasEnabled": {
                    "type": "boolean"
                },
                "atlasCoverageLayer": {
                    "type": "string"
                }
            },
            "required": [
                "name",
                "atlasEnabled"
            ],
            "type": "object"
        },
        "LayoutSummaries": {
            "additionalProperties": true,
            "properties": {
                "layouts": {
                    "items": {
                        "$ref": "#/definitions/LayoutSummary"
                    },
                    "type": "array"
                }
            },
            "required": [
                "layouts"
            ],
            "title": "LayoutSummaries",
            "type": "object"
        }
    },
...
    "info": {
        "title": "Lizmap-server-api",
        "description": "Lizmap server api for returning server and Qgis project's informations",
        "version": "1"
    }
}